requirements.txt
streamlit

numpy==1.26.4
pandas

opencv-python==4.6.0.66
pillow

easyocr
ultralytics
PyMuPDF

openpyxl
gdown

sqlalchemy
pyodbc

paddlepaddle==2.6.2
paddleocr==2.7.3